Remove obsolete views of old frontend
