--- cpmonitor/templates/administration_checklist.html
@@ -1,21 +1,0 @@
-{% extends "city-base.html" %}
-{% load static %}
-{% load martortags %}
-{% block title %}
-  LocalZero Monitoring - {{ city.name }}
-{% endblock title %}
-{% block text-content %}
-  <h1>Nachhaltigkeitsarchitektur in der Verwaltung</h1>
-  {% if city.assessment_administration %}<p>{{ city.assessment_administration | safe_markdown }}</p>{% endif %}
-{% endblock text-content %}
-{% block cards %}
-  <div class="card col-sm-12 col-lg-6">
-    <div class="card-body">
-      <div class="accordion accordion-flush" id="accordionFlushKAP">
-        {% for checklist_item in administration_checklist %}
-          {% include "snippets/checklist-item.html" %}
-        {% endfor %}
-      </div>
-    </div>
-  </div>
-{% endblock cards %}
