--- cpmonitor/templates/base.html
@@ -1,161 +1,0 @@
-{% load static %}
-{% load martortags %}
-<!DOCTYPE HTML>
-<html lang="de">
-  <head>
-    <meta charset="UTF-8">
-    <meta name="viewport"
-          content="width=device-width, initial-scale=1, viewport-fit=cover">
-    <meta name="description"
-          content="{% block description %}Monitoring von Klimaschutz-Maßnahmen in LocalZero-Kommunen{% endblock description %}">
-    <title>
-      {% block title %}
-        LocalZero Monitoring
-      {% endblock title %}
-    </title>
-    <link rel="stylesheet" href="{% static 'tabler/css/tabler.css' %}">
-    <link rel="stylesheet" href="{% static 'css/localzero.css' %}">
-    <link rel="stylesheet" href="{% static 'css/main.css' %}">
-    <script src="{% static 'jquery/jquery.min.js' %}"></script>
-    <script src="{% static 'tabler/js/tabler.min.js' %}" defer></script>
-    <script src="{% static 'apexcharts/apexcharts.min.js' %}"></script>
-  </head>
-  <body>
-    <div class="page">
-      <header class="lz-construction-banner">
-        <div class="container-xl">
-          Die Website befindet sich noch in Entwicklung <a href="https://github.com/GermanZero-de/klimaschutzmonitor">(Github-Projekt)</a>.
-        </div>
-      </header>
-      <header id="site-header">
-        <div class="site-header-inner">
-          {% block logo %}
-            <a class="site-branding" href="{% url 'index' %}" title="Home">
-              <div class="site-logo">&nbsp;</div>
-              <p class="site-title">LocalZero Monitoring</p>
-            </a>
-          {% endblock logo %}
-          <div id="site-menu">
-            <nav class="site-navigation">
-              <ul class="menu">
-                <li class="menu-item">
-                  <a href="{% url 'project' %}">Das Projekt</a>
-                </li>
-                <li class="menu-item">
-                  <a href="{% url 'ueber-uns' %}">Über uns</a>
-                </li>
-                <li class="menu-item" id="menu-item-toggleable">
-                  <a role="button">Kommunen</a>
-                  <ul class="sub-menu">
-                    {% for city in cities %}
-                      <li class="menu-item">
-                        <a href="{% url 'city' city.slug %}">{{ city.name }}</a>
-                      </li>
-                    {% endfor %}
-                  </ul>
-                </li>
-              </ul>
-            </nav>
-          </div>
-          <div class="nav-spenden-container">
-            <a class="btn nav-spenden"
-               href="https://localzero.net/jetzt-spenden"
-               target="_blank"
-               type="button">Jetzt spenden</a>
-          </div>
-          <div id="nav-burger" class="nav-burger">
-            <div>
-              <div class="hamburger"></div>
-              <div class="hamburger"></div>
-            </div>
-          </div>
-        </div>
-      </header>
-      {% block breadcrumbs %}
-        {% if breadcrumbs %}
-          <header class="lz-breadcrumbs-bar">
-            <div class="container-xl">
-              <ol class="breadcrumb breadcrumb-arrows" aria-label="breadcrumbs">
-                {% for breadcrumb in breadcrumbs %}
-                  <li class="breadcrumb-item {% if forloop.last %}active{% endif %}">
-                    <a href="{{ breadcrumb.url }}">{{ breadcrumb.label }}</a>
-                  </li>
-                {% endfor %}
-              </ol>
-            </div>
-          </header>
-        {% endif %}
-      {% endblock breadcrumbs %}
-      <div class="page-wrapper">
-        <div class="page-body">
-          {% block content %}
-            {% block text-content-container %}
-              <div class="container">
-                {% block text-content %}
-                {% endblock text-content %}
-              </div>
-            {% endblock text-content-container %}
-            {% block cards-container %}
-              <div class="container-fluid lz-cards-background">
-                <div class="container">
-                  <div class="space-y">
-                    {% block cards %}
-                    {% endblock cards %}
-                  </div>
-                </div>
-              </div>
-            {% endblock cards-container %}
-            <div class="container">
-              {% block more-text-content %}
-              {% endblock more-text-content %}
-            </div>
-          {% endblock content %}
-        </div>
-      </div>
-      <footer id="site-footer">
-        <div class="site-footer-inner">
-          <div>
-            <p>
-              Website und Software von LocalZero Monitoring sind unter der
-              <a href="https://github.com/GermanZero-de/klimaschutzmonitor/blob/main/LICENSE">MIT-Lizenz</a>
-              verfügbar.
-            </p>
-            <p xmlns
-               cc="https://creativecommons.org/ns#"
-               xmlns
-               dct="https://purl.org/dc/terms/">
-              <span property="dct:title">Monitoring-Inhalte</span>
-              von
-              <span property="cc:attributionName">LocalZero</span>
-              sind unter
-              {# djlint:off #}
-              {# djlint moves the end of the a tag to the next line, causing the following space to be highlighted on hover #}
-              <a href="https://creativecommons.org/licenses/by-sa/4.0/?ref=chooser-v1" target="_blank" rel="license noopener noreferrer">
-                CC BY-SA 4.0
-                <img class="cc-license-img" height="18" width="auto" alt="" src="{% static 'cc-logo.svg' %}">
-                <img class="cc-license-img" height="18" width="auto" alt="" src="{% static 'cc-by-logo.svg' %}">
-                <img class="cc-license-img" height="18" width="auto" alt="" src="{% static 'cc-sa-logo.svg' %}"></a>
-              {# djlint:on #}
-              verfügbar, wenn nicht anders angegeben.
-            </p>
-          </div>
-          <nav>
-            <ul>
-              <li>
-                <a href="https://github.com/GermanZero-de/klimaschutzmonitor">GitHub</a>
-              </li>
-              <li>
-                <a href="{% url 'impressum' %}">Impressum</a>
-              </li>
-              <li>
-                <a href="{% url 'datenschutz' %}">Datenschutz</a>
-              </li>
-            </ul>
-          </nav>
-        </div>
-      </footer>
-    </div>
-    <script src="{% static 'js/main.js' %}"></script>
-    <script src="{% static 'js/localzero.js' %}"></script>
-  </body>
-</html>
