--- cpmonitor/templates/cap_checklist.html
@@ -1,23 +1,0 @@
-{% extends "city-base.html" %}
-{% load static %}
-{% load martortags %}
-{% block title %}
-  LocalZero Monitoring - {{ city.name }}
-{% endblock title %}
-{% block text-content %}
-  <h1>Klimaaktionsplan (KAP)</h1>
-  {% if city.assessment_action_plan %}<p>{{ city.assessment_action_plan | safe_markdown }}</p>{% endif %}
-{% endblock text-content %}
-{% block cards %}
-  <div class="row">
-    <div class="card col-sm-12 col-lg-6">
-      <div class="card-body">
-        <div class="accordion accordion-flush" id="accordionFlushKAP">
-          {% for checklist_item in cap_checklist %}
-            {% include "snippets/checklist-item.html" %}
-          {% endfor %}
-        </div>
-      </div>
-    </div>
-  </div>
-{% endblock cards %}
