--- cpmonitor/templates/city-base.html
@@ -1,21 +1,0 @@
-{% extends "base.html" %}
-{% load martortags %}
-{% load static %}
-{% block logo %}
-  {% if local_group.logo %}
-    <a href="{% url 'city' city.slug %}" title="Home">
-      <img src="{{ local_group.logo.url }}"
-           width="auto"
-           height="60px"
-           alt="Logo der Lokalgruppe {{ local_group.name }}">
-    </a>
-  {% else %}
-    <a class="site-branding" href="{% url 'index' %}" title="Home">
-      <div class="site-logo">&nbsp;</div>
-      <p class="site-title">LocalZero Monitoring</p>
-    </a>
-  {% endif %}
-{% endblock logo %}
-{% block title %}
-  LocalZero Monitoring - {{ city.name }}
-{% endblock title %}
