--- cpmonitor/templates/datenschutz.html
@@ -1,17 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% load martortags %}
-{% block text-content %}
-  <header class="page-header">
-    <h1>Datenschutz</h1>
-  </header>
-  <p>
-    LocalZero gehört zum <a href="https://www.germanzero.de/" rel="noopener noreferrer nofollow">GermanZero e.V.</a>
-  </p>
-  <h3>Datenschutzerklärung</h3>
-  <p>
-    Es gelten die Datenschutzbestimmungen gemäß <a href="https://www.germanzero.de/datenschutz" target="_blank">https://www.germanzero.de/datenschutz</a>.
-  </p>
-{% endblock text-content %}
-{% block cards-container %}
-{% endblock cards-container %}
