--- cpmonitor/templates/energy_plan_checklist.html
@@ -1,130 +1,0 @@
-{% extends "city-base.html" %}
-{% load static %}
-{% load martortags %}
-{% block title %}
-  LocalZero Monitoring - {{ city.name }}
-{% endblock title %}
-{% block text-content %}
-  <h1>Wärmeplanung</h1>
-  <p>
-    Der Wärmewende kommt eine Schlüsselposition bei der Einhaltung des Paris Agreement und damit der Begrenzung der Klimaerhitzung auf unter 2 Grad zu.
-    Im Stromsektor gibt es etablierte Marktmechanismen, im Verkehrssektor sind die Lösungen bekannt, im Wärmesektor hingegen müssen Lösungen gefunden und Weichen gestellt werden.
-    Diese Lösungen müssen vor Ort verhandelt und geplant werden. Der Start ist eine fundierte Wärmeplanung. Zwei Punkte sind hierbei wichtig:
-    <ul>
-      <li>Qualitätsanspruch: Die Wärmeplanung muss dem Stand von Wissenschaft und Technik entsprechen.</li>
-      <li>
-        Ambitionsniveau: Mit Hilfe der Wärmeplanung die Dekarbonisierung der Wärmeversorgung bis 2035 abgeschlossen sein.
-      </li>
-    </ul>
-    Dafür stellt LocalZero die folgende Checkliste zur Verfügung.
-  </p>
-{% endblock text-content %}
-{% block cards %}
-  <div class="row">
-    <div class="card col-sm-12 col-lg-6">
-      <div class="card-body">
-        <div class="accordion accordion-flush" id="accordionFlushKAP">
-          <h2>1. Beschluss zur Durchführung</h2>
-          <p>Kommunaler Beschluss zur Durchführung der Wärmeplanung inkl. öffentlicher Bekanntmachung</p>
-          {% with checklist_item=energy_plan_checklist.0 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.1 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          <h2>2. Eignungsprüfung und verkürztes Verfahren (§ 14 WPG)</h2>
-          <p>
-            Frühzeitiges Ausschlussverfahren: Ausschluss von nicht geeigneten Quartieren/Gebieten für Wärmenetz oder Wasserstoffnetz (z.B. für ländliche, zersiedelte Räume)
-          </p>
-          {% with checklist_item=energy_plan_checklist.2 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.3 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          <h2>3. Bestandsaufnahme inkl. Wärmebedarfe Status quo (§ 15 WPG)</h2>
-          <p>Aktuelle Wärmeversorgung und Wärmebedarf feststellen.</p>
-          {% with checklist_item=energy_plan_checklist.4 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.5 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          <h2>4. Erstellung der Potenzialberechnungen (§ 16 WPG)</h2>
-          <p>
-            Ziel ist die Ausweisung von Wärmebedarf und Wärmeversorgung im Ist-Zustand und Ziel-Zustand, in Abstimmung mit jeweiligen Szenarien.
-            Hier sind zwei unterschiedliche Aspekte entscheidend:
-            <br />
-            (1) Identifizierung der Potenziale zur erneuerbaren Wärmeerzeugung
-            <br />
-            (2) Einschätzung über Potenziale zur Energieeinsparung sowie -effizienz durch Wärmebedarfsreduktion in Gebäuden sowie in industriellen oder gewerblichen Prozessen
-          </p>
-          {% with checklist_item=energy_plan_checklist.6 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.7 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          <h3>Kommunikation und Akteursbeteiligung</h3>
-          {% with checklist_item=energy_plan_checklist.8 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          <h2>5. Erstellung der Zielszenarien (§ 17 WPG)</h2>
-          <p>
-            <ul>
-              <li>Entwicklung des zukünftigen Wärmebedarfs</li>
-              <li>Flächenhafte Darstellung zur klimaneutralen Bedarfsdeckung mit jeweiligen Zwischenschritten</li>
-            </ul>
-          </p>
-          {% with checklist_item=energy_plan_checklist.9 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.10 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.11 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.12 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          <h3>Kommunikation und Akteursbeteiligung</h3>
-          {% with checklist_item=energy_plan_checklist.13 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          <h2>6. Einteilung in Wärmeversorgungsgebiete und -arten (§18 und 19 WPG)</h2>
-          <p>
-            Bei der Einteilung in Wärmeversorgungsgebiete passieren zwei Dinge:
-            <br />
-            (1) Einteilung in voraussichtliche Wärmeversorgungsgebiete und -arten
-            <br />
-            (2) Darstellung der Wärmeversorgungsarten für das Zieljahr
-            <br />
-            Aufteilung nach Wärmeversorgungsgebieten (Wärmenetz, Wasserstoff, dezentrale Gebiete), eingeteilt in wo welcher Gebietstyp (sehr) (un)wahrscheinlich bzw. (un)geeignet ist.
-            Aufteilung muss in Einklang mit vorliegendem/sich in der Erstellung befindlichen Wärmenetzbau- und -dekarbonisierungsfahrplan (Paragraph 32) sein
-          </p>
-          {% with checklist_item=energy_plan_checklist.14 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.15 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.16 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          <h2>7. Umsetzungsstrategie und konkrete Umsetzungsmaßnahmen (§ 20 WPG)</h2>
-          <p>
-            Ziel muss ein ambitionierter Transformationspfad mit klaren Maßnahmen und jahresscharfer Planung mit schnellstmöglichem Start sein. Die Kommune („planungsverantwortliche Stelle“) muss im Rahmen ihrer Möglichkeiten selbst Maßnahmen durchführen, mit denen das Zielszenario erreicht werden kann, oder Dritte dazu beauftragen.
-            Die Wärmeplanung ist eine Strategie, nicht einfach nur Daten und Fakten. Daher ist es wichtig, dass alle Akteure (s.u.) klare Rollen und Aufgaben haben und diese ineinandergreifen.
-          </p>
-          {% with checklist_item=energy_plan_checklist.17 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-          {% with checklist_item=energy_plan_checklist.18 %}
-            {% include "snippets/checklist-item.html" %}
-          {% endwith %}
-        </div>
-      </div>
-    </div>
-  </div>
-{% endblock cards %}
