--- cpmonitor/templates/impressum.html
@@ -1,70 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% load martortags %}
-{% block text-content %}
-  <header class="page-header">
-    <h1>Impressum</h1>
-  </header>
-  <p>
-    LocalZero gehört zum <a href="https://www.germanzero.de/" rel="noopener noreferrer nofollow">GermanZero e.V.</a>
-  </p>
-  <h3>Angaben gemäß § 5 TMG</h3>
-  <p>
-    GermanZero e.V.
-    <br>
-    Sitz des Vereins: Hamburg (Vereinsregisternummer 24224)
-    <br>
-    Geschäftsstelle Berlin
-    <br>
-    Franklinstraße 27, 10587 Berlin
-    <br>
-    vertreten durch den Vorstand: Sina Arndt, Anke Bytomski-Guerrier, Dr.-Ing. Claas Helmke, Peter Schwierz
-    E-Mail: <a href="mailto:info@germanzero.de" rel="noopener noreferrer nofollow">info@GermanZero.de</a>
-    Telefon: <a href="tel:03039807590" rel="noopener noreferrer nofollow">030 39807590</a>
-    Verantwortlicher i.S.v. § 55 Rundfunkstaatsvertrag (RStV): Dr. Julian Zuber
-    Steuernummer: 17/430/17389
-  </p>
-  <h3>Streitschlichtung</h3>
-  <p>
-    Wir sind nicht bereit oder verpflichtet, an Streitbeilegungsverfahren vor einer
-    Verbraucherschlichtungsstelle teilzunehmen.
-  </p>
-  <h3>Haftung für Inhalte</h3>
-  <p>
-    Als Diensteanbieter sind wir gemäß § 7 Abs.1 TMG für eigene Inhalte auf diesen Seiten nach den
-    allgemeinen Gesetzen verantwortlich. Nach §§ 8 bis 10 TMG sind wir als Diensteanbieter jedoch nicht
-    verpflichtet, übermittelte oder gespeicherte fremde Informationen zu überwachen oder nach Umständen zu
-    forschen, die auf eine rechtswidrige Tätigkeit hinweisen.
-    Verpflichtungen zur Entfernung oder Sperrung der Nutzung von Informationen nach den allgemeinen
-    Gesetzen bleiben hiervon unberührt. Eine diesbezügliche Haftung ist jedoch erst ab dem Zeitpunkt der
-    Kenntnis einer konkreten Rechtsverletzung möglich. Bei Bekanntwerden von entsprechenden
-    Rechtsverletzungen werden wir diese Inhalte umgehend entfernen.
-  </p>
-  <h3>Haftung für Links</h3>
-</p>
-<p>
-  Unser Angebot enthält Links zu externen Websites Dritter, auf deren Inhalte wir keinen Einfluss haben.
-  Deshalb können wir für diese fremden Inhalte auch keine Gewähr übernehmen. Für die Inhalte der
-  verlinkten Seiten ist stets der jeweilige Anbieter oder Betreiber der Seiten verantwortlich. Die
-  verlinkten Seiten wurden zum Zeitpunkt der Verlinkung auf mögliche Rechtsverstöße überprüft.
-  Rechtswidrige Inhalte waren zum Zeitpunkt der Verlinkung nicht erkennbar.
-  Eine permanente inhaltliche Kontrolle der verlinkten Seiten ist jedoch ohne konkrete Anhaltspunkte
-  einer Rechtsverletzung nicht zumutbar. Bei Bekanntwerden von Rechtsverletzungen werden wir derartige
-  Links umgehend entfernen.
-</p>
-<h3>Urheberrecht</h3>
-</p>
-<p>
-  Die durch die Seitenbetreiber erstellten Inhalte und Werke auf diesen Seiten unterliegen dem deutschen
-  Urheberrecht. Die Vervielfältigung, Bearbeitung, Verbreitung und jede Art der Verwertung außerhalb der
-  Grenzen des Urheberrechtes bedürfen der schriftlichen Zustimmung des jeweiligen Autors bzw.
-  Erstellers. Downloads und Kopien dieser Seite sind nur für den privaten, nicht kommerziellen Gebrauch
-  gestattet.
-  Soweit die Inhalte auf dieser Seite nicht vom Betreiber erstellt wurden, werden die Urheberrechte
-  Dritter beachtet. Insbesondere werden Inhalte Dritter als solche gekennzeichnet. Sollten Sie trotzdem
-  auf eine Urheberrechtsverletzung aufmerksam werden, bitten wir um einen entsprechenden Hinweis. Bei
-  Bekanntwerden von Rechtsverletzungen werden wir derartige Inhalte umgehend entfernen.
-</p>
-{% endblock text-content %}
-{% block cards-container %}
-{% endblock cards-container %}
