--- cpmonitor/templates/index.html
@@ -1,36 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% load martortags %}
-{% block text-content %}
-  <h1>Willkommen</h1>
-  <p>
-    <a href="https://localzero.net/">LocalZero</a> Monitoring ist eine Initiative von <a href="https://germanzero.de">GermanZero</a>, um mehr Transparenz
-    zum Fortschritt der Klimaneutralität deutscher Kommunen zu schaffen. Der Fortschritt wird von ehrenamtlichen Lokalteams in den
-    jeweiligen Kommunen regelmäßig aktualisiert.
-  </p>
-{% endblock text-content %}
-{% block cards %}
-  <div class="row row-deck row-cards">
-    {% for city in cities %}
-      <div class="col-sm-6 col-lg-3">
-        <a class="card" href="{% url 'city' city.slug %}">
-          <div class="card-body">
-            <div class="d-flex align-items-center">
-              <div class="h2 mb-3">{{ city.name }}</div>
-            </div>
-            <div class="lz-card-summary lines-3">{{ city.teaser }}</div>
-          </div>
-        </a>
-      </div>
-    {% endfor %}
-  </div>
-{% endblock cards %}
-{% block more-text-content %}
-  <h1>Mit Unterstützung von</h1>
-  <div class="pt-3">
-    <img alt="aid-five logo"
-         height="65"
-         src="{% static 'aidfive.jpg' %}"
-         width="auto">
-  </div>
-{% endblock more-text-content %}
