--- cpmonitor/templates/local-group.html
@@ -1,42 +1,0 @@
-{% extends "city-base.html" %}
-{% load static %}
-{% load martortags %}
-{% block title %}
-  LocalZero Monitoring - {{ city.name }}
-{% endblock title %}
-{% block text-content %}
-  <div class="row">
-    <div class="col">
-      <!-- Page  pre-title -->
-      <div class="page-pretitle">Lokalgruppe</div>
-      <h1>{{ local_group.name }}</h1>
-      <p>
-        <svg width="15" height="15">
-          <use href="{% static 'tabler-icons/tabler-sprite.svg' %}#tabler-link" />
-        </svg>
-        <a href="{{ local_group.website }}" target="_blank">{{ local_group.website }}</a>
-      </p>
-    </div>
-  </div>
-  <div class="row">
-    {% if local_group.featured_image %}
-      <div class="col-lg-5 col-md-6 col-sm-6">
-        <img src="{{ local_group.featured_image.url }}"
-             width="500"
-             height="300"
-             class="img-fluid"
-             alt="Bild der Lokalgruppe {{ local_group.name }}">
-      </div>
-    {% endif %}
-    <div class="{% if local_group.featured_image %}col-lg-7 col-md-6 col-sm-6{% else %}col{% endif %}">
-      {% if local_group.teaser %}
-        <p>
-          <strong>{{ local_group.teaser }}</strong>
-        </p>
-      {% endif %}
-      {{ local_group.description | safe_markdown | default:"Die Lokalgruppe hat keine Beschreibung eingegeben" }}
-    </div>
-  </div>
-{% endblock text-content %}
-{% block cards-container %}
-{% endblock cards-container %}
