--- cpmonitor/templates/project.html
@@ -1,17 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% load martortags %}
-{% block text-content %}
-  <h1>LocalZero Monitoring</h1>
-  <p>Städte und Kommunen leisten einen wichtigen Beitrag zur Klimaneutralität.</p>
-  <p>
-    LocalZero Monitoring ist eine Initiative von Localzero, um mehr Transparenz
-    zum Fortschritt der Klimaneutralität deutscher Kommunen zu schaffen. Der Fortschritt wird von ehrenamtlichen
-    Lokalteams in den jeweiligen Kommunen regelmäßig aktualisiert.
-  </p>
-  <p>
-    Du findest deine Kommune nicht? Dann <a href="https://localzero.net/mitmachen/lokal">finde Mitstreiter:innen</a> und gründe ein neues Team bei dir vor Ort!
-  </p>
-{% endblock text-content %}
-{% block cards-container %}
-{% endblock cards-container %}
