--- cpmonitor/templates/task.html
@@ -1,90 +1,0 @@
-{% extends "city-base.html" %}
-{% load martortags %}
-{% load static %}
-{% block title %}
-  LocalZero Monitoring - {{ city.name }}
-{% endblock title %}
-{% block text-content %}
-  <div class="row">
-    <div class="col">
-      <!-- Page  pre-title -->
-      <div class="page-pretitle">
-        {% if task.is_suggestion %}Vorgeschlagene{% endif %}
-        Maßnahme
-      </div>
-      {% if task.is_suggestion %}
-        <h1 class="text-secondary">{{ task.title }} [VORSCHLAG]</h1>
-      {% else %}
-        <h1>{{ task.title }}</h1>
-      {% endif %}
-    </div>
-  </div>
-  <div class="row">
-    <div class="col-sm-3">
-      <ul class="list-group mb-3 bg-light">
-        <li class="list-group-item d-flex justify-content-between lh-condensed">
-          <div>
-            <h4 class="my-0">Status</h4>
-          </div>
-          <span class="badge lz-status {{ task.get_execution_status_name }}">{{ task.get_execution_status_display }}</span>
-        </li>
-        <li class="list-group-item">
-          <div class="justify-content-between d-flex">
-            <h4 class="my-0">Beginn</h4>
-            <span class="text-muted">{{ task.actual_start|date:"d.m.Y"|default:"--" }}</span>
-          </div>
-          {% if task.planned_start %}
-            <div class="justify-content-between d-flex">
-              <span class="text-muted">geplant</span>
-              <span class="text-muted  {{ task.started_late|yesno:'lz-date-missed,' }}">{{ task.planned_start|date:"d.m.Y" }}</span>
-            </div>
-          {% endif %}
-        </li>
-        <li class="list-group-item">
-          <div class="justify-content-between d-flex">
-            <h4 class="my-0">Ende</h4>
-            <span class="text-muted">{{ task.actual_completion|date:"d.m.Y"|default:"--" }}</span>
-          </div>
-          {% if task.planned_completion %}
-            <div class="justify-content-between d-flex">
-              <span class="text-muted">geplant</span>
-              <span class="text-muted {{ task.completed_late|yesno:'lz-date-missed,' }}">{{ task.planned_completion|date:"d.m.Y" }}</span>
-            </div>
-          {% endif %}
-        </li>
-        <li class="list-group-item d-flex justify-content-between">
-          <div>
-            <h4 class="my-0">Zuständige Instanz</h4>
-          </div>
-          <span class="text-align-right">{{ task.responsible_organ | default:"--" }}</span>
-        </li>
-      </ul>
-    </div>
-    <div class="col-sm-9">
-      {% if task.teaser %}
-        <p>
-          <strong>{{ task.teaser }}</strong>
-        </p>
-      {% endif %}
-      {{ task.description | safe_markdown | default:"Diese Maßnahme hat keine Beschreibung" }}
-      {% if task.plan_assessment %}
-        <h3>Bewertung der geplanten Maßnahme</h3>
-        {{ task.plan_assessment | safe_markdown }}
-      {% endif %}
-      {% if task.execution_justification %}
-        <h3>Begründung Umsetzungsstand</h3>
-        {{ task.execution_justification | safe_markdown }}
-      {% endif %}
-      {% if task.responsible_organ_explanation %}
-        <h3>Zuständige Instanz</h3>
-        {{ task.responsible_organ_explanation | safe_markdown }}
-      {% endif %}
-      {% if task.supporting_ngos %}
-        <h3>Mit Unterstützung von</h3>
-        {{ task.supporting_ngos | safe_markdown }}
-      {% endif %}
-    </div>
-  </div>
-{% endblock text-content %}
-{% block cards-container %}
-{% endblock cards-container %}
