--- cpmonitor/templates/taskgroup.html
@@ -1,72 +1,0 @@
-{% extends "city-base.html" %}
-{% load static %}
-{% load martortags %}
-{% block title %}
-  LocalZero Monitoring - {{ city.name }}
-{% endblock title %}
-{% block text-content %}
-  {% if not node %}
-    <h1>Maßnahmen</h1>
-    {% if city.assessment_status %}
-      <p class="lz-description">
-        {{ city.assessment_status | safe_markdown }}
-      {% endif %}
-    </p>
-  {% endif %}
-  {% if node %}
-    <!-- Page pre-title -->
-    <div class="page-pretitle">Handlungsfeld</div>
-    <h1>{{ node.title }}</h1>
-    {% if node.teaser %}<p class="lz-teaser">{{ node.teaser }}</p>{% endif %}
-    <p class="lz-description">
-      {{ node.description | safe_markdown | default:"Dieses Handlungsfeld hat keine Beschreibung" }}
-    </p>
-  {% endif %}
-{% endblock text-content %}
-{% block cards %}
-  {% if tasks %}
-    <div class="row row-deck row-cards">
-      {% for task in tasks %}
-        <div class="col-sm-6 col-lg-3">{% include "snippets/taskcard.html" %}</div>
-      {% endfor %}
-    </div>
-  {% endif %}
-  {% if groups %}
-    {% if node or tasks %}
-      <div class="hr-text hr-text-left">
-        {% if node %}Weitere{% endif %}
-        Handlungsfelder
-      </div>
-    {% endif %}
-    <div class="row row-deck row-cards">
-      {% for group in groups %}
-        <div class="col-sm-6 col-lg-3">
-          <a href="{% url 'task' city.slug group.slugs %}"
-             class="card card-link card-link-pop">
-            <div class="card-body">
-              <div class="h3 mb-3">{{ group.title }}</div>
-              <div class="col">{{ group.subtasks_count }} Maßnahmen</div>
-              <div class="progress progress-lg progress-separated mb-3">
-                {% for value, status in group.status_proportions %}
-                  <div class="progress-bar lz-status {{ status.name }}"
-                       role="progressbar"
-                       data-value="{{ value }}"
-                       data-bs-toggle="tooltip"
-                       data-bs-placement="top"
-                       title="{{ value }} % {{ status.label }}"
-                       aria-label="{{ status.label }}">
-                  </div>
-                {% endfor %}
-              </div>
-            </div>
-            <div class="lz-card-details">
-              <svg width="20" height="20">
-                <use href="{% static 'tabler-icons/tabler-sprite.svg' %}#tabler-plus" />
-              </svg>
-            </div>
-          </a>
-        </div>
-      {% endfor %}
-    </div>
-  {% endif %}
-{% endblock cards %}
