--- cpmonitor/templates/ueber-uns.html
@@ -1,19 +1,0 @@
-{% extends "base.html" %}
-{% load static %}
-{% load martortags %}
-{% block text-content %}
-  <header class="page-header">
-    <h1>Über uns</h1>
-  </header>
-  <p>LocalZero Monitoring ist ein Produkt von LocalZero, der kommunalen Klimabewegung von GermanZero e.V.</p>
-  <p>
-    Wir machen Ort für Ort klimaneutral.
-    Diesen Weg gehen wir gemeinsam in kleinen, effizienten Teams und als großes Netzwerk. Wir sind hunderte aktive
-    Menschen, die sich für ein klimaneutrales Deutschland bis 2035 einsetzen.
-  </p>
-  <p>
-    Alle Details erfährst du unter <a href="https://localzero.net" rel="noopener noreferrer nofollow">LocalZero.net</a>.
-  </p>
-{% endblock text-content %}
-{% block cards-container %}
-{% endblock cards-container %}
