+++ cpmonitor/urls.py
@@ -11,10 +11,6 @@ from .views import SelectCityView, CapEditView, move_task
 prefix_kommune = ""
 
 urlpatterns = [
-    path("projekt/", views.project_view, name="project"),
-    path("impressum/", views.impressum_view, name="impressum"),
-    path("datenschutz/", views.datenschutz_view, name="datenschutz"),
-    path("ueber-uns/", views.ueber_uns_view, name="ueber-uns"),
     path(
         "favicon.ico",
         RedirectView.as_view(url=settings.STATIC_URL + "favicon.svg", permanent=True),
@@ -37,34 +33,6 @@ urlpatterns = [
         name="accept-invite",
     ),
     path("api/uploader/", views.markdown_uploader_view, name="markdown_uploader"),
-    path("", views.index_view, name="index"),
-    path(prefix_kommune + "<slug:city_slug>/", views.city_view, name="city"),
-    path(
-        prefix_kommune + "<slug:city_slug>/lokalgruppe/",
-        views.local_group_view,
-        name="local_group",
-    ),
-    path(
-        prefix_kommune + "<slug:city_slug>/kap_checkliste/",
-        views.cap_checklist_view,
-        name="cap_checklist",
-    ),
-    path(
-        prefix_kommune + "<slug:city_slug>/verwaltungsstrukturen_checkliste/",
-        views.administration_checklist_view,
-        name="administration_checklist",
-    ),
-    path(
-        prefix_kommune + "<slug:city_slug>/waermeplanung_checkliste/",
-        views.energy_plan_checklist_view,
-        name="energy_plan_checklist",
-    ),
-    path(prefix_kommune + "<slug:city_slug>/massnahmen/", views.task_view, name="task"),
-    path(
-        prefix_kommune + "<slug:city_slug>/massnahmen/<path:task_slugs>/",
-        views.task_view,
-        name="task",
-    ),
     #
     # REST API urls
     #
